+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. This projec
 
 ## 2.7.0-beta-1
 - Added support for Conversation API v0.1
+- Removed `nexmo.users.getConversations`
 
 ## 2.6.0-beta-1
 - Added beta2 API version for `get()` methods in `nexmo.users`, `nexmo.users.getConversations`, `nexmo.conversations`, `nexmo.conversations.members` and `nexmo.conversations.events`